Implementado token
Se implementa token del admin para todas las funciones menos las list

+++ api150/api/acts/delete.php
@@ -21,24 +21,35 @@
     $data = json_decode(file_get_contents("php://input"));
 
     $idRecibida = htmlspecialchars($_GET["idPrograma"]);
-    
-    if (!empty($idRecibida)) {
-        // Tengo todos los datos
-        // Comprobamos que la id corresponde a un registro 
-        if ($cf->comprobarExisteActoPorId($idRecibida)) {
-            // Efectivamente existe y se puede eliminar
-            $query = "DELETE FROM programas WHERE id_Programa like ".$idRecibida.";";
-            $stmt = $database->getConn()->prepare($query);
-            $stmt->execute();
-            echo json_encode(array(" error : 0, message : Elemento eliminado"));
-        } else {
-            //No existe y por lo tanto no se puede eliminar 
-            echo json_encode(array("error : 1, message : El registro no existe"));
-        }
+    $token = htmlspecialchars($_GET["token"]);
+
+    // Comprobamos que el token es de admin
+    if ($cf->comprobarTokenAdmin($token) == 1) {
+        // Token de admin
+        if (!empty($idRecibida)) {
+                // Tengo todos los datos
+                // Comprobamos que la id corresponde a un registro 
+                if ($cf->comprobarExisteActoPorId($idRecibida)) {
+                    // Efectivamente existe y se puede eliminar
+                    $query = "DELETE FROM programas WHERE id_Programa like ".$idRecibida.";";
+                    $stmt = $database->getConn()->prepare($query);
+                    $stmt->execute();
+                    echo json_encode(array(" error : 0, message : Elemento eliminado"));
+                } else {
+                    //No existe y por lo tanto no se puede eliminar 
+                    echo json_encode(array("error : 1, message : El registro no existe"));
+                }
+            } else {
+                echo json_encode("error : 1, message : faltan uno o más datos");
+            }
+
+
+    } elseif ($cf->comprobarTokenAdmin($token) == 0) {
+        echo json_encode("error : 2, message : no tiene permisos para realizar esta operación");
     } else {
-        echo json_encode("error : 1, message : faltan uno o más datos");
+        echo json_encode("error : 3, message : token no valido");
+
     }
     
-
 
 ?>
